Django/iCoder(Project)/iCoder/requirements.txt
pip-21.3.1
asgiref==3.4.1 
django==3.2.9 
pytz==2021.3 
sqlparse==0.4.2
Pillow-8.4.0
django-simple-search-1.0.2